requirements.txt
absl-py==0.15.0
aiosignal==1.3.1
akro==0.0.8
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cached-property==1.5.2
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
clang==5.0
click==8.1.3
cloudpickle==1.6.0
cma==2.7.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dm-tree==0.1.8
dowel==0.0.3
entrypoints==0.4
fastjsonschema==2.16.3
filelock==3.12.0
flatbuffers==1.12
frozenlist==1.3.3
garage==2021.3.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gym==0.18.3
h5py==3.1.0
idna==3.4
imageio==2.9.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==2.0.5
msgpack==1.0.5
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
ortools==9.3.10459
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyglet==1.5.15
Pygments==2.11.2
pyparsing==2.4.5
pyrsistent==0.19.3
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.10
PyYAML==5.1.2
pyzmq==22.3.0
qtconsole==5.4.2
QtPy==2.3.1
ray==2.4.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.2
setproctitle==1.3.2
six==1.15.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.4.1
tabulate==0.8.9
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.6.0
tensorflow-estimator==2.12.0
tensorflow-probability==0.13.0
termcolor==1.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.7.1
torchvision==0.4.1
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==3.7.4.3
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wincertstore==0.2
wrapt==1.12.1
zipp==3.15.0
